feat: tests succeed in multiple environemnts
