+++ .envrc
@@ -3,4 +3,4 @@ use asdf
 export PATH=$(npm bin):${PATH}
 export MOCK_PINNING_SERVER_PORT=3000
 export MOCK_PINNING_SERVER_SECRET=secret
-export DEBUG=1
+export DEBUG=0
