+++ CONTRIBUTING.md
@@ -1,0 +1,30 @@
+## @ipfs-shipyard/pinning-service-client@1.0.0
+
+This generator creates TypeScript/JavaScript client that utilizes fetch-api.
+
+### Building
+
+To build and compile the typescript sources to javascript use:
+```
+npm install
+npm run build
+```
+
+### Publishing
+
+First build the package then run ```npm publish```
+
+### Consuming
+
+navigate to the folder of your consuming project and run one of the following commands.
+
+_published:_
+
+```
+npm install @ipfs-shipyard/pinning-service-client@0.0.0 --save
+```
+
+_unPublished (not recommended):_
+
+```
+npm install PATH_TO_GENERATED_PACKAGE --save
