+++ test/browser.ts
@@ -1,17 +1,17 @@
-/* eslint-disable no-console */
+// /* eslint-disable no-console */
 
-import clientTests from './isomorphic-tests/client'
-// import configurationTests from './configuration.spec'
-import fetchPonyfill from 'fetch-ponyfill'
+// import clientTests from './isomorphic-tests/client'
+// // import configurationTests from './configuration.spec'
+// import fetchPonyfill from 'fetch-ponyfill'
 
-const setup = async () => {
-  return {
-    fetch: fetchPonyfill().fetch as GlobalFetch['fetch']
-  }
-}
+// const setup = async () => {
+//   return {
+//     fetch: fetchPonyfill().fetch as GlobalFetch['fetch']
+//   }
+// }
 
-// eslint-disable-next-line @typescript-eslint/no-misused-promises
-describe('browser', async (): Promise<void> => {
-  // await configurationTests(setup)
-  await clientTests(setup)
-})
+// // eslint-disable-next-line @typescript-eslint/no-misused-promises
+// describe('browser', async (): Promise<void> => {
+//   // await configurationTests(setup)
+//   await clientTests(setup)
+// })
